backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
pydantic==2.8.2

requests==2.32.3
httpx==0.27.2

beautifulsoup4==4.12.3
feedparser==6.0.11
PyYAML==6.0.2
